int para float - liberdade ao usuário

+++ Aulas/condicoes.c
@@ -1,14 +1,14 @@
 int main()
 {
 
-    int n1;
-    int n2;
+    float n1;
+    float n2;
 
     printf("Digite o primeiro número: ");
-    scanf("%d", &n1);
+    scanf("%f", &n1);
 
     printf("Digite o segundo número: ");
-    scanf("%d", &n2);
+    scanf("%f", &n2);
 
     if (n1 > n2)
     {
